Simplied code
Cleaned up hierarchy chain
